Initial Deploy of SQLite DB
The 'main.py' file runs the accompanying scripts which will create a local database (within project directory), print the SQLite version, create the DB tables and populate the tables with examples values. The database can be opened and viewed using DB Browser which acts as a user interface for SQLite. A unused script is included which can create the database in memory (RAM) rather than in the project directory.